requirements.txt
django==1.6.7
django-tastypie==0.12.0
lxml==3.4.0
defusedxml==0.4.1

